requirements.txt
Django>=5.2.7,<6.0.0
djangorestframework>=3.16.1,<4.0.0
djangorestframework-simplejwt>=5.4.0,<6.0.0
django-filter>=24.3,<25.0
django-cors-headers>=4.9.0,<5.0.0
django-prometheus>=2.4.1,<3.0.0
django-guardian>=3.2.0,<4.0.0
django-auditlog>=3.3.0,<4.0.0
drf-spectacular>=0.27.2,<0.28.0
psycopg2-binary>=2.9.11,<3.0.0
django-redis>=6.0.0,<7.0.0
celery>=5.4.0,<6.0.0
python-dotenv>=1.2.1,<2.0.0
pyotp>=2.9.0,<3.0.0
